enh: Rename spatial transformers
