+++ conda/environment.osx-64.lock
@@ -1,6 +1,6 @@
 # Generated by conda-lock.
 # platform: osx-64
-# input_hash: d85a4167f0d439e878aaa93f8928bcc04b3ee428f01bb40d4bf245626f80d69f
+# input_hash: cbc440a3b86cb94a853c3f25280dd51f8f51a1d42036620f15d72f6fa50ec531
 @EXPLICIT
 https://conda.anaconda.org/conda-forge/osx-64/bzip2-1.0.8-h0d85af4_4.tar.bz2#37edc4e6304ca87316e160f5ca0bd1b5
 https://conda.anaconda.org/conda-forge/osx-64/c-ares-1.19.1-h0dc2134_0.conda#b3e62631b4e1b9801477523ce1d6f355
@@ -227,6 +227,7 @@ https://conda.anaconda.org/conda-forge/osx-64/cffi-1.15.1-py310ha78151a_3.conda#
 https://conda.anaconda.org/conda-forge/noarch/comm-0.1.3-pyhd8ed1ab_0.conda#168ae0f82cdf7505048e81054c7354e4
 https://conda.anaconda.org/conda-forge/osx-64/curl-8.1.2-hbee3ae8_0.conda#6cc301a6c2ba26e29949818efdc133ca
 https://conda.anaconda.org/conda-forge/noarch/dacite-1.8.0-pyhd8ed1ab_0.conda#3696792ba70ab3374320fe6041a82286
+https://conda.anaconda.org/conda-forge/noarch/deprecation-2.1.0-pyh9f0ad1d_0.tar.bz2#7b6747d7cc2076341029cff659669e8b
 https://conda.anaconda.org/conda-forge/osx-64/ffmpeg-5.1.2-gpl_h8b4fe81_106.conda#3c62afa3457aeb6f058b40f1f41f772a
 https://conda.anaconda.org/conda-forge/noarch/flake8-6.0.0-pyhd8ed1ab_0.conda#e9345ba05d71742412b8aa6992ad9457
 https://conda.anaconda.org/conda-forge/osx-64/fonttools-4.40.0-py310h6729b98_0.conda#15dcc729f2102b4b64e0e6f53f2fb377
@@ -317,7 +318,7 @@ https://conda.anaconda.org/conda-forge/noarch/prompt_toolkit-3.0.38-hd8ed1ab_0.c
 https://conda.anaconda.org/conda-forge/osx-64/qt-main-5.15.8-h1d3b3f8_6.conda#5d816352174169f7ab45fb54a0fba4ed
 https://conda.anaconda.org/conda-forge/noarch/urllib3-1.26.15-pyhd8ed1ab_0.conda#27db656619a55d727eaf5a6ece3d2fd6
 https://conda.anaconda.org/conda-forge/noarch/wslink-1.11.0-pyhd8ed1ab_0.conda#70f4b52c0e703936aa1812d8b14a5bb2
-https://conda.anaconda.org/conda-forge/noarch/botocore-1.29.157-pyhd8ed1ab_0.conda#a4f0350e313a7705f917565d768005a4
+https://conda.anaconda.org/conda-forge/noarch/botocore-1.29.158-pyhd8ed1ab_0.conda#ccfde1c6ce8dbb4af5590c88a0482248
 https://conda.anaconda.org/conda-forge/noarch/ipython-8.14.0-pyhd1c38e8_0.conda#f56fab4cea853c2248105b6cd7d79bf0
 https://conda.anaconda.org/conda-forge/osx-64/matplotlib-3.7.1-py310h2ec42d9_0.conda#958029c1236746d27149fca648c41cab
 https://conda.anaconda.org/conda-forge/noarch/nbclient-0.7.4-pyhd8ed1ab_0.conda#f7aa15f77d29b11caa1df1eb15383c59
@@ -335,7 +336,7 @@ https://conda.anaconda.org/conda-forge/noarch/s3transfer-0.6.1-pyhd8ed1ab_0.cond
 https://conda.anaconda.org/conda-forge/noarch/sphinx-4.5.0-pyh6c4a22f_0.tar.bz2#46b38d88c4270ff9ba78a89c83c66345
 https://conda.anaconda.org/conda-forge/noarch/torchinfo-1.8.0-pyhd8ed1ab_0.conda#fba98fc95a945e2a6b93a77b39dd52dc
 https://conda.anaconda.org/conda-forge/noarch/torchmetrics-0.11.4-pyhd8ed1ab_0.conda#480cb2b6d502003e937d9e4326bc398f
-https://conda.anaconda.org/conda-forge/noarch/boto3-1.26.157-pyhd8ed1ab_0.conda#e834ba9d0ed83f1b51b3f7e98673a3d2
+https://conda.anaconda.org/conda-forge/noarch/boto3-1.26.158-pyhd8ed1ab_0.conda#2e39307a3987518e49eb30ca51e3cfb5
 https://conda.anaconda.org/conda-forge/noarch/google-auth-oauthlib-1.0.0-pyhd8ed1ab_0.conda#5ba9d95acb9add67f468ecdee6595c0f
 https://conda.anaconda.org/conda-forge/noarch/ipywidgets-8.0.6-pyhd8ed1ab_0.conda#68627a08556e4a273e4c7bfc84251457
 https://conda.anaconda.org/conda-forge/noarch/jupyter_console-6.6.3-pyhd8ed1ab_0.conda#7cf6f52a66f8e3cd9d8b6c231262dcab
