+++ pyproject.toml
@@ -40,6 +40,7 @@ dynamic = ["version"]
 
 dependencies = [
     "dacite",
+    "deprecation",
     "pandas",
     "pyyaml",
     "SimpleITK>=2.0",
