+++ src/deepali/version.py
@@ -1,0 +1,4 @@
+from importlib import metadata
+
+
+__version__ = metadata.version("deepali-hf")
